weblog_receiver/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-cloud-pubsub>=2.19.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.5.0
